Fix util::lookup_table initialization
The third and last argument in util::lookup_table's constructor is `highest_item`,
which specifies up to what index should the table be default initialized for.

It did not however do that, and instead it would just allocate up to that item
and then initialize up to the highest number provided by the initializer list.

It now correctly initializes the table as intended.

Signed-off-by: Vinícius Schütz Piva <vinicius.vsczpv@outlook.com>